[WIP] Invites feature

+++ src/firebase/firebase-list.js
@@ -1,6 +1,4 @@
-import { firebaseDb } from './firebase';
-
-
+import { firebaseDb } from "./firebase";
 /*
  Represent a Firebase collection.
  If given a rootPath and rootDocId then this might be under a sub collection.
@@ -10,7 +8,15 @@ import { firebaseDb } from './firebase';
  path = tasks
  */
 export class FirebaseList {
-  constructor(actions, modelClass, path = null, query = null, orderBy = null, rootPath = null, rootDocId = null) {
+  constructor(
+    actions,
+    modelClass,
+    path = null,
+    query = null,
+    orderBy = null,
+    rootPath = null,
+    rootDocId = null
+  ) {
     this._actions = actions;
     this._modelClass = modelClass;
     this._path = path;
@@ -61,15 +67,38 @@ export class FirebaseList {
   }
 
   get collectionPath() {
-    if(this._rootPath && this._rootDocId) {
-      return firebaseDb.collection(this._rootPath).doc(this._rootDocId).collection(this._path);
-    }else {
+    if (this._rootPath && this._rootDocId) {
+      return firebaseDb
+        .collection(this._rootPath)
+        .doc(this._rootDocId)
+        .collection(this._path);
+    } else {
       return firebaseDb.collection(`${this._path}`);
     }
   }
 
+  cleanObjectBeforeSend(obj) {
+    const stringified = JSON.stringify(obj);
+
+    // We need to parse string back to object and return it
+    const parsed = JSON.parse(stringified);
+
+    return parsed;
+  }
+
   push(value) {
     return this.collectionPath.add(value);
+  }
+
+  pushBatch(values) {
+    const batch = firebaseDb.batch();
+
+    values.forEach(value => {
+      const ref = this.collectionPath.doc();
+      batch.set(ref, value);
+    });
+
+    return batch.commit();
   }
 
   remove(id) {
@@ -77,49 +106,60 @@ export class FirebaseList {
   }
 
   set(id, value) {
-    return this.collectionPath.doc(id).set(value);
+    return this.collectionPath.doc(id).set(this.cleanObjectBeforeSend(value));
   }
 
   update(id, value) {
-    return this.collectionPath.doc(id).update(value);
+    return this.collectionPath
+      .doc(id)
+      .update(this.cleanObjectBeforeSend(value));
   }
 
   subscribe(emit) {
     // This list might be under a root collection
     let collection = this.collectionPath;
-    if(this._query) {
+    if (this._query) {
       collection = collection.where(
-        this._query[0],this._query[1],this._query[2]);
+        this._query[0],
+        this._query[1],
+        this._query[2]
+      );
     }
-    if(this._orderBy) {
-      collection = collection.orderBy(this._orderBy.name,
-        this._orderBy.direction);
+    if (this._orderBy) {
+      collection = collection.orderBy(
+        this._orderBy.name,
+        this._orderBy.direction
+      );
     }
     let initialized = false;
     let list = [];
 
     let unsubscribe = collection.onSnapshot(snapshot => {
-      if(!initialized) {
+      if (!initialized) {
         emit(this._actions.onLoad(snapshot.docs));
         initialized = true;
         return;
       }
       const isLocalChange = snapshot.metadata.hasPendingWrites;
       snapshot.docChanges().forEach(change => {
-          if (change.type === "added") {
-            if (initialized) {
-              emit(this._actions.onAdd(this.unwrapSnapshot(change.doc), isLocalChange));
-            }
-            else {
-              list.push(this.unwrapSnapshot(change.doc));
-            }
+        if (change.type === "added") {
+          if (initialized) {
+            emit(
+              this._actions.onAdd(
+                this.unwrapSnapshot(change.doc),
+                isLocalChange
+              )
+            );
+          } else {
+            list.push(this.unwrapSnapshot(change.doc));
           }
-          if (change.type === "modified") {
-              emit(this._actions.onChange(this.unwrapSnapshot(change.doc)));
-          }
-          if (change.type === "removed") {
-              emit(this._actions.onRemove(this.unwrapSnapshot(change.doc)));
-          }
+        }
+        if (change.type === "modified") {
+          this._actions.onChange && emit(this._actions.onChange(this.unwrapSnapshot(change.doc)));
+        }
+        if (change.type === "removed") {
+          this._actions.onRemove && emit(this._actions.onRemove(this.unwrapSnapshot(change.doc)));
+        }
       });
     });
 
@@ -127,7 +167,7 @@ export class FirebaseList {
   }
 
   unsubscribe() {
-    if(this._unsubscribe) {
+    if (this._unsubscribe) {
       this._unsubscribe();
     }
   }
@@ -144,7 +184,12 @@ export class FirebaseList {
  We are adding the id and a function that allow to perform object.get
  */
 export function firebaseCollectionToList(collection) {
-  return collection.map(document => {
-    return Object.assign(document.data(),{get: (object)=>document[object], id: document.id});
-  });
+  if (collection) {
+    return collection.map(document => {
+      return Object.assign(document.data(), {
+        get: object => document[object],
+        id: document.id
+      });
+    });
+  }
 }
